+++ towhee/dag/dataframe_repr.py
@@ -20,7 +20,7 @@ import yaml
 import requests
 import os
 
-from towhee.base_repr import BaseRepr
+from towhee.dag.base_repr import BaseRepr
 from towhee.dag.variable_repr import VariableRepr
